Fix missing `.set noreorder` et al
